[fix] tee: add sec_world_id to struct tee_shm
Change the type of sec_world_id member of struct tee_shm to u64 to be
able to store the 64-bit Global Handle defined in FF-A.

+++ include/linux/tee_drv.h
@@ -196,7 +196,7 @@ struct tee_shm {
 	struct dma_buf *dmabuf;
 	u32 flags;
 	int id;
-	u32 sec_world_id;
+	u64 sec_world_id;
 };
 
 /**
